+++ lib/pycall/truffleruby/tuple.rb
@@ -3,12 +3,8 @@ require "pycall/truffleruby/pyenum"
 module PyCall
   class Tuple < PyEnumerable
 
-    def initialize(foreign)
-      if Truffle::Interop.foreign?(foreign)
-        super foreign
-      else
-        super PyCall.builtins.tuple(*foreign)
-      end
+    def initialize(*args)
+      super PyCall.builtins.tuple(args).__pyptr__#TODO: do without wrap->unwrap->wrap
     end
   end
 end
