+++ spec/pycall/libpython/finder_spec.rb
@@ -4,20 +4,22 @@ require 'tmpdir'
 module PyCall
   module LibPython
     ::RSpec.describe Finder do
-      describe '.find_python_config' do
-        context 'when the given python command is not python' do
-          specify do
-            expect {
-              Finder.find_python_config('echo')
-            }.to raise_error(PyCall::PythonNotFound)
+      if RUBY_ENGINE != "truffleruby"
+        describe '.find_python_config' do
+          context 'when the given python command is not python' do
+            specify do
+              expect {
+                Finder.find_python_config('echo')
+              }.to raise_error(PyCall::PythonNotFound)
+            end
           end
         end
-      end
 
-      describe '.investigate_python_config' do
-        subject { Finder.investigate_python_config('python') }
+        describe '.investigate_python_config' do
+          subject { Finder.investigate_python_config('python') }
 
-        it { is_expected.to be_a(Hash) }
+          it { is_expected.to be_a(Hash) }
+        end
       end
     end
   end
